requirements.txt
# Core PyTorch
torch>=2.0.0

# DeepSeek / Transformers
transformers>=4.36.0
accelerate

# LoRA
peft>=0.7.0

# Sionna for CSI data generation
tensorflow>=2.12.0
sionna

# Utilities
numpy>=1.24.0
pyyaml
h5py
